Change Default Style For Presentation
Last night, we had a practice session for CPP on Sea. One of the
comments from the meeting was to use a light theme for the live
presentation. This commit sets the default theme to 'League' and
changes the syntax highlighting theme to 'github'. I have also
included the 'xcode' syntax highlighting theme to provide an
alternative (if required).